Move must.go into metric.go
